// libs/begetter/src/schematics/domain/schema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface DomainOptions {
  /**
   * Domain name
   */
  name: string;
  /**
   * app name
   */
  app?: string;
  /**
   * Is this feature module lazy loaded?
   */
  lazy?: boolean;
  /**
   * runtime platform
   */
  platform?: 'web' | 'mobile' | 'desktop' | 'nest';
  [k: string]: any;
}